EntityClassInspector 리팩토링
